Spotify/requirements.txt
Django>=4.2  
djangorestframework>=3.14.0  
djangorestframework-simplejwt>=5.3.0  
django-allauth>=0.60.0  
dj-rest-auth>=4.0.1  
django-cors-headers>=4.3.0  
social-auth-app-django>=5.2.0  
psycopg2-binary>=2.9.9  
gunicorn>=21.2.0  
Pillow>=8.4.0 
whitenoise>=6.6.0  
